+++ backend/src/main/java/com/expensetracker/main/domain/income/dto/TotalIncomeAmountDto.java
@@ -1,5 +1,7 @@
 package com.expensetracker.main.domain.income.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -8,15 +10,15 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class TotalIncomeAmountDto {
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private Long incomeGroupId;
 
-    public TotalIncomeAmountDto(Double totalAmount) {
+    public TotalIncomeAmountDto(BigDecimal totalAmount) {
         this.totalAmount = totalAmount;
         this.incomeGroupId = null;
     }
 
-    public TotalIncomeAmountDto(Double totalAmount, Long incomeGroupId) {
+    public TotalIncomeAmountDto(BigDecimal totalAmount, Long incomeGroupId) {
         this.totalAmount = totalAmount;
         this.incomeGroupId = incomeGroupId;
     }
